test: add a test for preserving marks
