+++ cypress/integration/rich-text/RichTextEditor.spec.ts
@@ -643,163 +643,6 @@ describe('Rich Text Editor', { viewportHeight: 2000 }, () => {
     }
   });
 
-  describe('Lists', () => {
-    const lists = [
-      {
-        getList: () => richText.toolbar.ul,
-        listType: BLOCKS.UL_LIST,
-        label: 'Unordered List (UL)',
-      },
-      { getList: () => richText.toolbar.ol, listType: BLOCKS.OL_LIST, label: 'Ordered List (OL)' },
-    ];
-
-    lists.forEach((test) => {
-      describe(test.label, () => {
-        it('should be visible', () => {
-          test.getList().should('be.visible');
-        });
-
-        it('should add a new list', () => {
-          richText.editor.click();
-
-          test.getList().click();
-          // TODO: Find a way to test deeper lists
-          /*
-            Having issues with `.type('{enter})` to break lines.
-            The error is:
-            Cannot resolve a Slate node from DOM node: [object HTMLSpanElement]
-          */
-          richText.editor.type('item 1');
-
-          const expectedValue = doc(
-            block(
-              test.listType,
-              {},
-              block(BLOCKS.LIST_ITEM, {}, block(BLOCKS.PARAGRAPH, {}, text('item 1', [])))
-            ),
-            emptyParagraph()
-          );
-
-          richText.expectValue(expectedValue);
-        });
-
-        it('should untoggle the list', () => {
-          richText.editor.click();
-
-          test.getList().click();
-
-          richText.editor.type('some text');
-
-          test.getList().click();
-
-          const expectedValue = doc(
-            block(BLOCKS.PARAGRAPH, {}, text('some text', [])),
-            emptyParagraph()
-          );
-
-          richText.expectValue(expectedValue);
-        });
-
-        it('should unwrap blockquote', () => {
-          addBlockquote('some text');
-
-          test.getList().click();
-
-          const expectedValue = doc(
-            block(
-              test.listType,
-              {},
-              block(BLOCKS.LIST_ITEM, {}, block(BLOCKS.PARAGRAPH, {}, text('some text', [])))
-            ),
-            emptyParagraph()
-          );
-
-          richText.expectValue(expectedValue);
-        });
-
-        it('should allow heading as direct child of <li>', () => {
-          richText.editor.click();
-          test.getList().click();
-
-          richText.toolbar.toggleHeading(BLOCKS.HEADING_1);
-          richText.editor.type('heading');
-
-          const expectedValue = doc(
-            block(
-              test.listType,
-              {},
-              block(BLOCKS.LIST_ITEM, {}, block(BLOCKS.HEADING_1, {}, text('heading')))
-            ),
-            emptyParagraph()
-          );
-
-          richText.expectValue(expectedValue);
-        });
-
-        it('should allow HR as direct child of <li>', () => {
-          richText.editor.click();
-          test.getList().click();
-
-          richText.toolbar.hr.click();
-
-          const expectedValue = doc(
-            block(test.listType, {}, block(BLOCKS.LIST_ITEM, {}, block(BLOCKS.HR, {}))),
-            emptyParagraph()
-          );
-
-          richText.expectValue(expectedValue);
-        });
-
-        it('should allow embedded entry as direct child of <li>', () => {
-          richText.editor.click();
-          test.getList().click();
-
-          richText.toolbar.embed('entry-block');
-
-          const expectedValue = doc(
-            block(test.listType, {}, block(BLOCKS.LIST_ITEM, {}, entryBlock())),
-            emptyParagraph()
-          );
-
-          richText.expectValue(expectedValue);
-        });
-
-        it('should allow embedded asset as direct child of <li>', () => {
-          richText.editor.click();
-          test.getList().click();
-
-          richText.toolbar.embed('asset-block');
-
-          const expectedValue = doc(
-            block(test.listType, {}, block(BLOCKS.LIST_ITEM, {}, assetBlock())),
-            emptyParagraph()
-          );
-
-          richText.expectValue(expectedValue);
-        });
-
-        it('should allow block quotes as direct child of <li>', () => {
-          richText.editor.click();
-          test.getList().click();
-
-          richText.toolbar.quote.click();
-          richText.editor.type('quote');
-
-          const expectedValue = doc(
-            block(
-              test.listType,
-              {},
-              block(BLOCKS.LIST_ITEM, {}, block(BLOCKS.QUOTE, {}, paragraphWithText('quote')))
-            ),
-            emptyParagraph()
-          );
-
-          richText.expectValue(expectedValue);
-        });
-      });
-    });
-  });
-
   describe('New Line', () => {
     it('should add a new line on a paragraph', () => {
       richText.editor
